Restructure Lambda functions with consistent naming and directory pattern
- Rename ParseJobAttributesFunction to ParseJobPropertiesFunction
- Move Lambda code to src/BatchJobsStates/ParseJobProperties/app.py
- Follow consistent directory pattern: [ServiceName]/[FunctionName]/app.py
- Add ParseJobPropertiesFunction resource to template.yaml with proper configuration
- Update state machine to reference new function name
- Add Lambda invoke permissions to state machine policies
- Update all test imports and documentation references
- Verify SAM build and template validation pass successfully
