+++ tests/BatchJobStatesTests/test_parse_job_attributes.py
@@ -1,18 +1,18 @@
 #!/usr/bin/env python3
 """
-Test script for ParseJobAttributesFunction Lambda function
+Test script for ParseJobPropertiesFunction Lambda function
 """
 
 import json
 import sys
 import os
 
-# Add the BatchJobsStates directory to Python path to import the function
+# Add the ParseJobProperties directory to Python path to import the function
 script_dir = os.path.dirname(os.path.abspath(__file__))
-batch_jobs_states_dir = os.path.join(os.path.dirname(os.path.dirname(script_dir)), "src", "BatchJobsStates")
-sys.path.insert(0, batch_jobs_states_dir)
+parse_job_properties_dir = os.path.join(os.path.dirname(os.path.dirname(script_dir)), "src", "BatchJobsStates", "ParseJobProperties")
+sys.path.insert(0, parse_job_properties_dir)
 
-from ParseJobAttributesFunction import lambda_handler
+from app import lambda_handler
 
 
 def load_test_event(filename):
@@ -135,7 +135,7 @@ def test_error_handling():
 
 
 if __name__ == "__main__":
-    print("Running ParseJobAttributesFunction tests...\n")
+    print("Running ParseJobPropertiesFunction tests...\n")
     
     try:
         test_starting_event()
